Restructured the visualization module.
