--- cocotools/src/visualize/bbox.rs
@@ -1,11 +1,0 @@
-use crate::annotations::coco;
-use image;
-use imageproc::{drawing::draw_hollow_rect_mut, rect::Rect};
-
-#[allow(clippy::cast_sign_loss, clippy::cast_possible_truncation)]
-pub fn draw_bbox(img: &mut image::RgbImage, bbox: &coco::Bbox, color: image::Rgb<u8>) {
-    let rect =
-        Rect::at(bbox.left as i32, bbox.top as i32).of_size(bbox.width as u32, bbox.height as u32);
-
-    draw_hollow_rect_mut(img, rect, color);
-}
